requirements.txt
pybaseball>=2.2.5
MLB-StatsAPI>=1.7.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.30.0
plotly>=5.18.0
